+++ packages/components/src/font-size-picker/index.js
@@ -33,6 +33,8 @@ function FontSizePicker(
 		fontSizes = [],
 		disableCustomFontSizes = false,
 		onChange,
+		/** @type {'default' | '__unstable-large'} */
+		size = 'default',
 		value,
 		withSlider = false,
 		withReset = true,
@@ -53,7 +55,7 @@ function FontSizePicker(
 	 * than six and a select control when they are more.
 	 */
 	const fontSizesContainComplexValues = fontSizes.some(
-		( { size } ) => ! isSimpleCssValue( size )
+		( { size: sizeArg } ) => ! isSimpleCssValue( sizeArg )
 	);
 	const shouldUseSelectControl = fontSizes.length > 5;
 	const options = useMemo(
@@ -168,6 +170,7 @@ function FontSizePicker(
 					shouldUseSelectControl &&
 					! showCustomValueControl && (
 						<CustomSelectControl
+							__nextUnconstrainedWidth
 							className={ `${ baseClassName }__select` }
 							label={ __( 'Font size' ) }
 							hideLabelFromVision
@@ -186,6 +189,7 @@ function FontSizePicker(
 									setShowCustomValueControl( true );
 								}
 							} }
+							size={ size }
 						/>
 					) }
 				{ ! shouldUseSelectControl && ! showCustomValueControl && (
@@ -199,6 +203,7 @@ function FontSizePicker(
 							);
 						} }
 						isBlock
+						size={ size }
 					>
 						{ options.map( ( option ) => (
 							<ToggleGroupControlOption
@@ -238,6 +243,7 @@ function FontSizePicker(
 											);
 										}
 									} }
+									size={ size }
 									units={ hasUnits ? units : [] }
 								/>
 							</FlexItem>